grep: add color.grep.branch
The configuration option color.grep.branch can be used to specify the
highlighting color for tree names.

Currently, it is displayed using the color.grep.filename as the both
values are in the same buffer. This buffer is not splitted in two
strings in the grep_source object. There is no impact on the raw output.

Signed-off-by: Thomas Faivre <thomas.faivre@6wind.com>